+++ qmake/library/proitems.h
@@ -155,8 +155,8 @@ public:
     int toInt(bool *ok = nullptr, int base = 10) const { return toQStringView().toInt(ok, base); }
     short toShort(bool *ok = nullptr, int base = 10) const { return toQStringView().toShort(ok, base); }
 
-    uint hash() const { return m_hash; }
-    static uint hash(const QChar *p, int n);
+    size_t hash() const { return m_hash; }
+    static size_t hash(const QChar *p, int n);
 
     ALWAYS_INLINE QStringView toQStringView() const { return QStringView(m_string).mid(m_offset, m_length); }
 
@@ -184,8 +184,8 @@ private:
     QString m_string;
     int m_offset, m_length;
     int m_file;
-    mutable uint m_hash;
-    uint updatedHash() const;
+    mutable size_t m_hash;
+    size_t updatedHash() const;
     friend size_t qHash(const ProString &str);
     friend QString operator+(const ProString &one, const ProString &two);
     friend class ProKey;
